made some aestethic changes
