+++ src/forms/LoginForm.py
@@ -3,8 +3,8 @@ from wtforms import StringField, PasswordField, validators
 
 
 class LoginValidation(FlaskForm):  # FlaskForm verwenden
-    user_name_pid = StringField('', [validators.DataRequired()],
+    user_name = StringField('', [validators.DataRequired()],
                                 render_kw={'autofocus': True, 'placeholder': 'Enter User'})
 
-    user_pid_Password = PasswordField('', [validators.DataRequired()],
+    user_password = PasswordField('', [validators.DataRequired()],
                                       render_kw={'autofocus': True, 'placeholder': 'Enter your login Password'})
