+++ src/templates/login.html
@@ -34,17 +34,17 @@
                                     </div>
 
                                     <form method="POST" class="form-horizontal needs-validation"
-                                          action="{{ url_for('index') }}">
+                                          action="{{ url_for('users_blueprint.index') }}">
                                         {{ form.csrf_token }}
                                         <div class="form-group row">
                                             <div class="col-sm-10">
-                                                {{ render_field(form.user_name_pid, class_="form-control", value="") }}
+                                                {{ render_field(form.user_name, class_="form-control", value="") }}
                                             </div>
                                         </div>
 
                                         <div class="form-group row">
                                             <div class="col-sm-10">
-                                                {{ render_field(form.user_pid_Password , class_="form-control",
+                                                {{ render_field(form.user_password , class_="form-control",
                                                 value="") }}
                                             </div>
                                         </div>
